+++ packages/sdk/CHANGELOG.md
@@ -2,6 +2,21 @@
 
 ## [UNRELEASED]
 
+## [1.3.0] - 2022-11-03
+
+- feat: add to request dapps by blockchainType property
+- fix yarn.lock
+
+## [1.2.1] - 2022-11-01
+
+- chore: expose dapp id in sdk api
+
+## [1.2.0] - 2022-10-26
+
+- feature: add api for add/remove members in group threads
+- feature: add api for rename group threads
+- chore: switch dialect rest api to v2
+
 ## [1.1.0] - 2022-10-05
 
 - feature: add read/unread messages api for dapp notification feed